+++ apps/web/src/app/[locale]/(profile)/[username]/_components/dashboard/challenge-history-columns.tsx
