+++ apps/web/src/app/[locale]/(profile)/[username]/completed/page.tsx
@@ -1,3 +1,5 @@
+import { getServerAuthSession } from '@repo/auth/server';
+import { prisma } from '@repo/db';
 import {
   Card,
   CardContent,
@@ -5,10 +7,11 @@ import {
   CardHeader,
   CardTitle,
 } from '@repo/ui/components/card';
-import { CompletedTab } from '../_components/dashboard/completed-tab';
 import { notFound } from 'next/navigation';
-import { prisma } from '@repo/db';
-import { getServerAuthSession } from '@repo/auth/server';
+import { createCompletedSubmissionCacheKey } from '~/app/[locale]/challenge/[slug]/submissions/[[...catchAll]]/cache-keys';
+import { withUnstableCache } from '~/utils/withUnstableCache';
+import { getChallengeHistoryByCategory } from '../_components/dashboard/_actions';
+import ChallengeHistory from '../_components/dashboard/challenge-history';
 
 interface Props {
   params: {
@@ -35,6 +38,13 @@ export default async function CompletedPage({ params: { username: usernameFromQu
 
   if (!user) return notFound();
 
+  const challenges = await withUnstableCache({
+    fn: getChallengeHistoryByCategory,
+    args: ['completed', user.id],
+    keys: [`completed-challenges-${user.id}`],
+    tags: [createCompletedSubmissionCacheKey(user.id)],
+  });
+
   return (
     <Card className="col-span-4 md:min-h-[calc(100vh_-_56px_-_6rem)]">
       <CardHeader>
@@ -44,7 +54,7 @@ export default async function CompletedPage({ params: { username: usernameFromQu
         </CardDescription>
       </CardHeader>
       <CardContent>
-        <CompletedTab userId={user.id} />
+        <ChallengeHistory challenges={challenges} />
       </CardContent>
     </Card>
   );
