+++ apps/web/src/app/[locale]/(profile)/[username]/shared-solutions/_components/shared-solutions-columns.tsx
