--- apps/web/src/components/Search/index.tsx
@@ -1,132 +1,0 @@
-import { useRouter } from 'next/router';
-import { createContext, useCallback, useContext, useEffect, useRef, useState } from 'react';
-import { createPortal } from 'react-dom';
-import { useActionKey } from '~/utils/use-action-key';
-
-const SearchContext = createContext({});
-
-export function SearchProvider({ children }) {
-  const router = useRouter();
-  const [isOpen, setIsOpen] = useState(false);
-  const [initialQuery, setInitialQuery] = useState(null);
-
-  const onOpen = useCallback(() => {
-    setIsOpen(true);
-  }, [setIsOpen]);
-
-  const onClose = useCallback(() => {
-    setIsOpen(false);
-  }, [setIsOpen]);
-
-  const onInput = useCallback(
-    (e) => {
-      setIsOpen(true);
-      setInitialQuery(e.key);
-    },
-    [setIsOpen, setInitialQuery],
-  );
-
-  useDocSearchKeyboardEvents({
-    isOpen,
-    onOpen,
-    onClose,
-  });
-
-  useEffect(() => {
-    // Prepend "Components" to Tailwind UI results that are shown in the "recent" view
-    if (!isOpen) {
-      try {
-        const data = JSON.parse(window.localStorage.getItem(key));
-        window.localStorage.setItem(key, JSON.stringify(data));
-      } catch {}
-    }
-  }, [isOpen]);
-
-  return (
-    <>
-      <SearchContext.Provider
-        value={{
-          isOpen,
-          onOpen,
-          onClose,
-          onInput,
-        }}
-      >
-        {children}
-      </SearchContext.Provider>
-      {isOpen ? createPortal(
-          <div>
-            <Spotlight />
-          </div>,
-          document.body,
-        ) : null}
-    </>
-  );
-}
-
-export function SearchButton({ children, ...props }) {
-  const searchButtonRef = useRef<HTMLButtonElement | null>(null);
-  const actionKey = useActionKey();
-  const { onOpen, onInput } = useContext(SearchContext);
-
-  useEffect(() => {
-    function onKeyDown(event: KeyboardEvent) {
-      if (searchButtonRef && searchButtonRef.current === document.activeElement && onInput) {
-        if (/[a-zA-Z0-9]/.test(String.fromCharCode(event.code))) {
-          onInput(event);
-        }
-      }
-    }
-    window.addEventListener('keydown', onKeyDown);
-    return () => {
-      window.removeEventListener('keydown', onKeyDown);
-    };
-  }, [onInput, searchButtonRef]);
-
-  return (
-    <button type="button" ref={searchButtonRef} onClick={onOpen} {...props}>
-      {typeof children === 'function' ? children({ actionKey }) : children}
-    </button>
-  );
-}
-
-function useDocSearchKeyboardEvents({
-  isOpen,
-  onOpen,
-  onClose,
-}: {
-  isOpen: boolean;
-  onOpen: () => void;
-  onClose: () => void;
-}) {
-  useEffect(() => {
-    function onKeyDown(event: KeyboardEvent) {
-      function open() {
-        // We check that no other DocSearch modal is showing before opening
-        // another one.
-        if (!document.body.classList.contains('DocSearch--active')) {
-          onOpen();
-        }
-      }
-
-      if (
-        (event.keyCode === 27 && isOpen) ||
-        (event.key === 'k' && (event.metaKey || event.ctrlKey)) ||
-        (!isEditingContent(event) && event.key === '/' && !isOpen)
-      ) {
-        event.preventDefault();
-
-        if (isOpen) {
-          onClose();
-        } else if (!document.body.classList.contains('DocSearch--active')) {
-          open();
-        }
-      }
-    }
-
-    window.addEventListener('keydown', onKeyDown);
-    return () => {
-      window.removeEventListener('keydown', onKeyDown);
-    };
-  }, [isOpen, onOpen, onClose]);
-}
